Accept mentor advices for module1
