+++ src/view/filter/filter-tpl.js
@@ -1,5 +1,3 @@
-import { createElement } from '../render.js';
-
 const createNewFilterTemplate = () => (
   `<form class="trip-filters" action="#" method="get">
   <div class="trip-filters__filter">
@@ -21,20 +19,4 @@ const createNewFilterTemplate = () => (
 </form>`
 );
 
-export default class FilterView {
-  getTemplate() {
-    return createNewFilterTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
+export default createNewFilterTemplate;
