+++ src/view/point-list/point-list-tpl.js
@@ -1,0 +1,3 @@
+const createPointListTemplate = () => '<ul class="trip-events__list"></ul>';
+
+export default createPointListTemplate;
