+++ src/view/sort/sort-tpl.js
@@ -1,5 +1,3 @@
-import {createElement} from '../render.js';
-
 const createSortTemplate = () => (
   `          <form class="trip-events__trip-sort  trip-sort" action="#" method="get">
   <div class="trip-sort__item  trip-sort__item--day">
@@ -29,20 +27,4 @@ const createSortTemplate = () => (
 </form>`
 );
 
-export default class SortView {
-  getTemplate() {
-    return createSortTemplate();
-  }
-
-  getElement() {
-    if (!this.element) {
-      this.element = createElement(this.getTemplate());
-    }
-
-    return this.element;
-  }
-
-  removeElement() {
-    this.element = null;
-  }
-}
+export default createSortTemplate;
