+++ src/view/sort/sort-view.js
@@ -1,0 +1,20 @@
+import {createElement} from '../../render.js';
+import createSortTemplate from './sort-tpl.js';
+
+export default class SortView {
+  getTemplate() {
+    return createSortTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
